refactor(scripts): consolidate, DRY, and npm-ify all user-facing scripts; update docs and references
