+++ docker-compose.yml
@@ -59,7 +59,7 @@ services:
     command: >
       sh -c "
         sleep 10 &&
-        npx hardhat run scripts/setup_dev.js --network hardhat_docker &&
+                  npx hardhat run scripts/deploy.js --network hardhat_docker &&
         echo 'Contract deployment complete'
       "
     depends_on:
