+++ docs/COMMANDS.md
@@ -2,7 +2,28 @@
 
 > **Single Source of Truth**: Common development commands used across the project.
 
-## Quick Start Commands
+## 📋 Available npm Scripts
+
+Run `npm run` to see all available scripts. Here are the most commonly used:
+
+```bash
+# Testing
+npm test                    # Run smart contract tests
+npm run test:ci-local       # Test CI-like conditions locally
+npm run test:integration    # Run integration tests with setup
+
+# Deployment
+npm run deploy:local        # Deploy to local Hardhat network
+npm run deploy              # Deploy to configured network
+
+# Code Quality
+npm run format              # Format all code files
+npm run format:check        # Check code formatting
+npm run lint:sol            # Lint Solidity contracts
+npm run lint:sol:fix        # Fix Solidity linting issues
+```
+
+## 🚀 Quick Start Commands
 
 ```bash
 # Initial setup
@@ -10,47 +31,57 @@ npm install && poetry install
 cp .env.example .env
 
 # Start development stack
-npx hardhat node                    # Terminal 1: Blockchain
-npx hardhat run scripts/deploy.js   # Terminal 2: Deploy contracts
+npx hardhat node              # Terminal 1: Blockchain
+npm run deploy:local          # Terminal 2: Deploy contracts
 poetry run uvicorn src.api.main:app --reload  # Terminal 3: API server
 ```
 
-## Blockchain Commands
+## 🔧 Blockchain Commands
 
+### Smart Contract Development
 ```bash
-# Contract Development
+# Contract compilation and testing
 npx hardhat compile                 # Compile contracts
-npx hardhat test                   # Run Solidity tests
+npm test                           # Run Solidity tests (alias: npx hardhat test)
 npx hardhat coverage              # Generate coverage report
 npx hardhat console --network localhost  # Interactive console
 
-# Deployment
-npx hardhat run scripts/deploy.js --network localhost
-npx hardhat verify --network <network> <address>
-
-# Development Node
-npx hardhat node                   # Start local blockchain
+# Contract deployment
+npm run deploy:local               # Deploy to localhost (recommended)
+npm run deploy                     # Deploy to default network
+npx hardhat run scripts/deploy.js --network <network>  # Deploy to specific network
+npx hardhat verify --network <network> <address>      # Verify on block explorer
 ```
 
-## Backend Commands
-
+### Development Blockchain
 ```bash
-# Development Server
+npx hardhat node                   # Start local blockchain
+npx hardhat node --hostname 0.0.0.0  # Start with external access
+```
+
+## 🐍 Backend Commands
+
+### Development Server
+```bash
 poetry run uvicorn src.api.main:app --reload    # Start API server
 poetry run uvicorn src.api.main:app --reload --port 8001  # Custom port
+```
 
-# Testing
+### Testing & Quality
+```bash
+# Python testing
 poetry run pytest                  # Run Python tests
 poetry run pytest --cov=src       # With coverage
 poetry run pytest -v              # Verbose output
+npm run test:integration           # Run integration tests (recommended)
 
-# Code Quality
+# Code quality
 poetry run black src tests         # Format code
 poetry run ruff check src tests    # Lint code
 poetry run mypy src                # Type checking
 ```
 
-## Docker Commands
+## 🐳 Docker Commands
 
 ```bash
 # Container Management (Docker Compose V2)
@@ -64,24 +95,29 @@ docker compose logs hardhat        # View blockchain logs
 # Replace 'docker compose' with 'docker-compose' in above commands
 ```
 
-## Quality Assurance
+## 🔍 Quality Assurance
 
 ```bash
-# Pre-commit Setup
+# Automated testing and checks
+npm run test:ci-local              # Test CI-like conditions locally (recommended)
+./scripts/test_ci_locally.sh       # Direct script execution
+
+# Pre-commit setup
 poetry run pre-commit install      # Install hooks
 poetry run pre-commit run --all-files  # Run all checks
 
-# CI Testing (Local)
-npm run test:ci-local              # Test CI-like conditions locally
-./scripts/test_ci_locally.sh       # Direct script execution
+# Code formatting and linting
+npm run format                     # Format all files (recommended)
+npm run format:check               # Check formatting
+npm run lint:sol                   # Lint Solidity (recommended)
+npm run lint:sol:fix               # Fix Solidity issues (recommended)
 
-# Manual Quality Checks
+# Manual quality checks
 npm test && poetry run pytest      # Run all tests
-npx solhint contracts/**/*.sol     # Lint Solidity
 npx hardhat test --gas-reporter    # Gas usage analysis
 ```
 
-## Environment Management
+## 📦 Environment Management
 
 ```bash
 # Dependencies
@@ -94,7 +130,42 @@ npm install <package>              # Add Node.js package
 node --version && python --version  # Check versions
 poetry env info                    # Poetry environment details
 npx hardhat --version              # Hardhat version
+npm run                            # List all available npm scripts
 ```
+
+## 🛠️ Advanced Usage
+
+### Script Analysis
+```bash
+# Script-specific commands (when npm scripts aren't sufficient)
+npx hardhat run scripts/deploy.js --network mainnet  # Deploy to mainnet
+./scripts/run_integration_tests.sh                   # Direct integration test execution
+./scripts/test_ci_locally.sh                         # Direct CI test execution
+
+# Note: scripts/_common.sh is a utility and should not be run directly
+```
+
+### Development Workflow
+```bash
+# Typical development cycle
+npm run test:ci-local              # 1. Test everything locally
+npm run deploy:local               # 2. Deploy contracts
+npm run test:integration           # 3. Run integration tests
+npm run format && npm run lint:sol # 4. Format and lint code
+```
+
+---
+
+## 🎯 Recommended Usage Patterns
+
+| Task | Recommended Command | Alternative |
+|------|-------------------|-------------|
+| **Deploy locally** | `npm run deploy:local` | `npx hardhat run scripts/deploy.js --network localhost` |
+| **Test contracts** | `npm test` | `npx hardhat test` |
+| **Integration tests** | `npm run test:integration` | `./scripts/run_integration_tests.sh` |
+| **Local CI testing** | `npm run test:ci-local` | `./scripts/test_ci_locally.sh` |
+| **Format code** | `npm run format` | `npx prettier --write "**/*.{js,json,sol,yml,yaml}"` |
+| **Lint Solidity** | `npm run lint:sol` | `npx solhint "contracts/**/*.sol"` |
 
 ---
