+++ scripts/run_integration_tests.sh
@@ -3,45 +3,33 @@
 
 set -e
 
+# Source common utilities
+source "$(dirname "$0")/_common.sh"
+
 echo "🧪 Running TicketChain Integration Tests"
 echo "========================================"
 
-# Colors for output
-GREEN='\033[0;32m'
-RED='\033[0;31m'
-YELLOW='\033[1;33m'
-NC='\033[0m' # No Color
-
 # Check if Hardhat node is running
-if ! curl -s http://localhost:8545 > /dev/null; then
-    echo -e "${RED}❌ Hardhat node is not running!${NC}"
-    echo "Please start it with: npx hardhat node"
-    exit 1
-fi
-echo -e "${GREEN}✅ Hardhat node is running${NC}"
+check_hardhat_node || exit 1
 
 # Check if contract is deployed
 if [ -z "$TICKET_CONTRACT_ADDRESS" ] && [ ! -f ".env" ]; then
     echo -e "${YELLOW}⚠️  Contract not deployed. Deploying now...${NC}"
     npx hardhat run scripts/deploy.js --network localhost
 
-    # Source the .env file to get contract address
-    if [ -f ".env" ]; then
-        source .env
-        echo -e "${GREEN}✅ Contract deployed to: $TICKET_CONTRACT_ADDRESS${NC}"
-    fi
+    # Load the newly created .env file
+    load_env_file && echo -e "${GREEN}✅ Contract deployed to: $TICKET_CONTRACT_ADDRESS${NC}"
 elif [ -f ".env" ] && [ -z "$TICKET_CONTRACT_ADDRESS" ]; then
     # Load from .env if available
-    source .env
-    echo -e "${GREEN}✅ Using existing contract at: $TICKET_CONTRACT_ADDRESS${NC}"
+    load_env_file && echo -e "${GREEN}✅ Using existing contract at: $TICKET_CONTRACT_ADDRESS${NC}"
 fi
 
 # Check if API is running
-if ! curl -s http://localhost:8000/api/v1/health > /dev/null; then
+if ! check_api_health; then
     echo -e "${YELLOW}⚠️  API server is not running. Starting it...${NC}"
 
-    # Source .env to ensure the API gets the contract address
-    source .env
+    # Load .env to ensure the API gets the contract address
+    load_env_file
     echo -e "${GREEN}📋 Starting API with contract address: $TICKET_CONTRACT_ADDRESS${NC}"
 
     # Export the variable so it's available to the subprocess
@@ -50,9 +38,7 @@ if ! curl -s http://localhost:8000/api/v1/health > /dev/null; then
     API_PID=$!
 
     # Wait for API to start
-    sleep 5
-
-    if ! curl -s http://localhost:8000/api/v1/health > /dev/null; then
+    if ! wait_for_service "API server" "check_api_health" 30 1; then
         echo -e "${RED}❌ Failed to start API server${NC}"
         exit 1
     fi
